Create Pima_Indians_Diabetes
The goal of the Pima Indians Dataset was to predict whether or not a patient has diabetes. The project involved exploring and visualizing the dataset to understand the distribution and relationships among variables. To apply various machine-learning models to predict diabetes and evaluate their performance. Identify the best-performing model based on accuracy, precision, recall, and F1-score.